Moves native-to-lifted piping code from MCToIRLifter to Lift.cpp
