+++ include/anvill/Lift.h
@@ -42,13 +42,6 @@ namespace anvill {
 class Program;
 struct ValueDecl;
 
-std::string CreateFunctionName(uint64_t addr);
-std::string CreateVariableName(uint64_t addr);
-
-// Adapt `src` to another type (likely an integer type) that is `dest_type`.
-llvm::Value *AdaptToType(llvm::IRBuilder<> &ir, llvm::Value *src,
-                         llvm::Type *dest_type);
-
 // Produce one or more instructions in `in_block` to load and return
 // the lifted value associated with `decl`.
 llvm::Value *LoadLiftedValue(const ValueDecl &decl,
